+++ AssetTracking/Computer.cs
@@ -4,7 +4,7 @@ public class Computer : Asset
 {
     public override string Type => "Computer";
 
-    public Computer(string brand, string model, int price, Office office, DateTime purchaseDate) : base(brand, model, price, office, purchaseDate)
+    public Computer(string brand, string model, double price, Office office, DateTime purchaseDate) : base(brand, model, price, office, purchaseDate)
     {
     }
 }
